refactor(engine): consolidate runtime editor tooling and plan updates

+++ zircon_editor/src/ui/retained_host/callback_dispatch/template_bridge/workbench/transform_edit.rs
@@ -78,7 +78,7 @@ impl BuiltinWorkbenchWindowTemplateSurfaceBridge {
                 "entity://selected",
                 [InspectorFieldChange::new(
                     field_id,
-                    UiBindingValue::Float(scalar),
+                    UiBindingValue::Float(f64::from(scalar)),
                 )],
             ),
         )))
